memrowset: fix on-stack MRSRow crash when row size is greater than 256 bytes
This fixes a dumb bug where we accidentally used uint8_t to store the size
of an MRS row. For wide rows, this overflowed and caused a crash.

This doesn't add a test, since we already have KUDU-665 filed to add tests
for these sorts of cases.

Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7074
Tested-by: jenkins

+++ src/kudu/tablet/memrowset.h
@@ -142,7 +142,7 @@ struct MSBTreeTraits : public btree::BTreeTraits {
 // plus a single row of the given schema, then constructs an MRSRow object which
 // points into that stack storage.
 #define DEFINE_MRSROW_ON_STACK(memrowset, varname, slice_name) \
-  uint8_t varname##_size = sizeof(MRSRow::Header) + \
+  size_t varname##_size = sizeof(MRSRow::Header) + \
                            ContiguousRowHelper::row_size((memrowset)->schema_nonvirtual()); \
   uint8_t varname##_storage[varname##_size]; \
   Slice slice_name(varname##_storage, varname##_size); \
